requirements.txt
boto3==1.19.7
python-dotenv==0.19.2
Jinja2==3.0.2
PyYAML==6.0
xmltodict==0.12.0
pandas==1.3.1
openpyxl
requests
tqdm


